ZCS-5821 Making HAB a network feature
