+++ store/src/java/com/zimbra/cs/account/ZAttrDomain.java
@@ -14402,26 +14402,26 @@ public abstract class ZAttrDomain extends NamedEntry {
     }
 
     /**
-     * domain level root of hierarchical address book
+     * root group id of hierarchical address book
      *
      * @return zimbraHierarchicalAddressBookRoot, or empty array if unset
      *
      * @since ZCS 8.8.10
      */
-    @ZAttr(id=3028)
+    @ZAttr(id=3072)
     public String[] getHierarchicalAddressBookRoot() {
         return getMultiAttr(Provisioning.A_zimbraHierarchicalAddressBookRoot, true, true);
     }
 
     /**
-     * domain level root of hierarchical address book
+     * root group id of hierarchical address book
      *
      * @param zimbraHierarchicalAddressBookRoot new value
      * @throws com.zimbra.common.service.ServiceException if error during update
      *
      * @since ZCS 8.8.10
      */
-    @ZAttr(id=3028)
+    @ZAttr(id=3072)
     public void setHierarchicalAddressBookRoot(String[] zimbraHierarchicalAddressBookRoot) throws com.zimbra.common.service.ServiceException {
         HashMap<String,Object> attrs = new HashMap<String,Object>();
         attrs.put(Provisioning.A_zimbraHierarchicalAddressBookRoot, zimbraHierarchicalAddressBookRoot);
@@ -14429,7 +14429,7 @@ public abstract class ZAttrDomain extends NamedEntry {
     }
 
     /**
-     * domain level root of hierarchical address book
+     * root group id of hierarchical address book
      *
      * @param zimbraHierarchicalAddressBookRoot new value
      * @param attrs existing map to populate, or null to create a new map
@@ -14437,7 +14437,7 @@ public abstract class ZAttrDomain extends NamedEntry {
      *
      * @since ZCS 8.8.10
      */
-    @ZAttr(id=3028)
+    @ZAttr(id=3072)
     public Map<String,Object> setHierarchicalAddressBookRoot(String[] zimbraHierarchicalAddressBookRoot, Map<String,Object> attrs) {
         if (attrs == null) attrs = new HashMap<String,Object>();
         attrs.put(Provisioning.A_zimbraHierarchicalAddressBookRoot, zimbraHierarchicalAddressBookRoot);
@@ -14445,14 +14445,14 @@ public abstract class ZAttrDomain extends NamedEntry {
     }
 
     /**
-     * domain level root of hierarchical address book
+     * root group id of hierarchical address book
      *
      * @param zimbraHierarchicalAddressBookRoot new to add to existing values
      * @throws com.zimbra.common.service.ServiceException if error during update
      *
      * @since ZCS 8.8.10
      */
-    @ZAttr(id=3028)
+    @ZAttr(id=3072)
     public void addHierarchicalAddressBookRoot(String zimbraHierarchicalAddressBookRoot) throws com.zimbra.common.service.ServiceException {
         HashMap<String,Object> attrs = new HashMap<String,Object>();
         StringUtil.addToMultiMap(attrs, "+" + Provisioning.A_zimbraHierarchicalAddressBookRoot, zimbraHierarchicalAddressBookRoot);
@@ -14460,7 +14460,7 @@ public abstract class ZAttrDomain extends NamedEntry {
     }
 
     /**
-     * domain level root of hierarchical address book
+     * root group id of hierarchical address book
      *
      * @param zimbraHierarchicalAddressBookRoot new to add to existing values
      * @param attrs existing map to populate, or null to create a new map
@@ -14468,7 +14468,7 @@ public abstract class ZAttrDomain extends NamedEntry {
      *
      * @since ZCS 8.8.10
      */
-    @ZAttr(id=3028)
+    @ZAttr(id=3072)
     public Map<String,Object> addHierarchicalAddressBookRoot(String zimbraHierarchicalAddressBookRoot, Map<String,Object> attrs) {
         if (attrs == null) attrs = new HashMap<String,Object>();
         StringUtil.addToMultiMap(attrs, "+" + Provisioning.A_zimbraHierarchicalAddressBookRoot, zimbraHierarchicalAddressBookRoot);
@@ -14476,14 +14476,14 @@ public abstract class ZAttrDomain extends NamedEntry {
     }
 
     /**
-     * domain level root of hierarchical address book
+     * root group id of hierarchical address book
      *
      * @param zimbraHierarchicalAddressBookRoot existing value to remove
      * @throws com.zimbra.common.service.ServiceException if error during update
      *
      * @since ZCS 8.8.10
      */
-    @ZAttr(id=3028)
+    @ZAttr(id=3072)
     public void removeHierarchicalAddressBookRoot(String zimbraHierarchicalAddressBookRoot) throws com.zimbra.common.service.ServiceException {
         HashMap<String,Object> attrs = new HashMap<String,Object>();
         StringUtil.addToMultiMap(attrs, "-" + Provisioning.A_zimbraHierarchicalAddressBookRoot, zimbraHierarchicalAddressBookRoot);
@@ -14491,7 +14491,7 @@ public abstract class ZAttrDomain extends NamedEntry {
     }
 
     /**
-     * domain level root of hierarchical address book
+     * root group id of hierarchical address book
      *
      * @param zimbraHierarchicalAddressBookRoot existing value to remove
      * @param attrs existing map to populate, or null to create a new map
@@ -14499,7 +14499,7 @@ public abstract class ZAttrDomain extends NamedEntry {
      *
      * @since ZCS 8.8.10
      */
-    @ZAttr(id=3028)
+    @ZAttr(id=3072)
     public Map<String,Object> removeHierarchicalAddressBookRoot(String zimbraHierarchicalAddressBookRoot, Map<String,Object> attrs) {
         if (attrs == null) attrs = new HashMap<String,Object>();
         StringUtil.addToMultiMap(attrs, "-" + Provisioning.A_zimbraHierarchicalAddressBookRoot, zimbraHierarchicalAddressBookRoot);
@@ -14507,13 +14507,13 @@ public abstract class ZAttrDomain extends NamedEntry {
     }
 
     /**
-     * domain level root of hierarchical address book
+     * root group id of hierarchical address book
      *
      * @throws com.zimbra.common.service.ServiceException if error during update
      *
      * @since ZCS 8.8.10
      */
-    @ZAttr(id=3028)
+    @ZAttr(id=3072)
     public void unsetHierarchicalAddressBookRoot() throws com.zimbra.common.service.ServiceException {
         HashMap<String,Object> attrs = new HashMap<String,Object>();
         attrs.put(Provisioning.A_zimbraHierarchicalAddressBookRoot, "");
@@ -14521,14 +14521,14 @@ public abstract class ZAttrDomain extends NamedEntry {
     }
 
     /**
-     * domain level root of hierarchical address book
+     * root group id of hierarchical address book
      *
      * @param attrs existing map to populate, or null to create a new map
      * @return populated map to pass into Provisioning.modifyAttrs
      *
      * @since ZCS 8.8.10
      */
-    @ZAttr(id=3028)
+    @ZAttr(id=3072)
     public Map<String,Object> unsetHierarchicalAddressBookRoot(Map<String,Object> attrs) {
         if (attrs == null) attrs = new HashMap<String,Object>();
         attrs.put(Provisioning.A_zimbraHierarchicalAddressBookRoot, "");
